feat(contentunderstanding): port CLI helper tests + fix changelog path
* Fix stale tools-dir path in CHANGELOG (still said .github/skills/_shared/
  after the earlier move to sdk/contentunderstanding/tools/cu-skill/).

* Refactor: extract pure helpers from CreateAndTestCommand and
  CreateAndTestRouterCommand into .Helpers.cs partial-class files so
  they can be linked into the package's test project without dragging
  in Azure.Identity / cu-skill's HintPath dependency. The originals
  keep the Azure.* entry-point code; the .Helpers.cs files are pure
  System.Text.Json.Nodes (no Azure.* imports), mirroring the
  SchemaValidator.cs pattern we already had.

* New SkillCreateAndTestCommandTests + SkillCreateAndTestRouterCommandTests
  covering the 6 pure-helper tests Python ships that translate to .NET:
  Summarize leaf-row flattening, SummarizeRouted per-category
  denominator + zero-fill, WireInnerIds missing-alias + extra-inner +
  prebuilt-passthrough (these last 3 work in .NET because our
  WireInnerIds returns a tuple of (Patched, Errors), unlike Java/JS),
  ParseInnerArg parse + missing-equals error.

  The 2 Python tests we still don't cover are the CLI integration tests
  for --help and 'invalid-schema-exits-before-client-built' — both
  require monkey-patching the Azure client, which isn't idiomatic in C#.

* Widen visibility of SummarizeRouted, WireInnerIds, and ParseInnerArg
  from private to internal so the test project can call them.

All 8 new tests pass on net10.0 (26/26 Skill* total).
